+++ packages/snap-preact-demo/tests/cypress/e2e/templates/recommendationBundleList/recommendationBundleList.cy.js
@@ -152,10 +152,12 @@ describe('BundledRecommendations', () => {
 		});
 
 		it('can click on a result and go to that page', function () {
-			cy.document().then((doc) => {
-				cy.snapController(config?.selectors?.recommendation.controller).then(({ store }) => {
-					cy.get(config?.selectors?.recommendation.result).should('exist');
-					let url = doc.querySelector(`${config?.selectors?.recommendation.result} a`).attributes?.href?.value;
+			cy.get(config?.selectors?.recommendation.result).should('exist');
+			cy.get(`${config?.selectors?.recommendation.result} a`)
+				.first()
+				.invoke('attr', 'href')
+				.then((url) => {
+					expect(url).to.be.a('string').and.not.be.empty;
 					cy.get(`${config?.selectors?.recommendation.result} a`)
 						.first()
 						.click({ force: true })
@@ -163,7 +165,6 @@ describe('BundledRecommendations', () => {
 							cy.location('href').should('include', url);
 						});
 				});
-			});
 		});
 
 		describe('Tests Custom Result Component', () => {