+++ packages/snap-preact/components/src/themes/themeComponents/searchHorizontal.ts
@@ -18,17 +18,20 @@ export const searchHorizontalThemeComponentProps: ThemeComponent<'searchHorizont
 		'searchHorizontal facetsHorizontal': {
 			limit: 9,
 		},
-		'searchHorizontal mobileSidebar': {
+		'searchHorizontal sidebar': {
 			layout: ['filterSummary', 'facets', 'banner.left'],
 		},
 	},
 
 	mobile: {
+		searchHorizontal: {
+			hideSidebar: false,
+		},
 		'searchHorizontal toolbar.top': {
 			layout: [['banner.header'], ['searchHeader', '_']],
 		},
 		'searchHorizontal toolbar.middle': {
-			layout: [['paginationInfo', '_'], ['mobileSidebar', '_', 'sortBy'], ['banner.banner']],
+			layout: [['paginationInfo', '_'], ['button.sidebar-toggle', '_', 'sortBy'], ['banner.banner']],
 		},
 		'searchHorizontal results': {
 			columns: 2,