+++ packages/snap-preact/src/Templates/Stores/LibraryStore.ts
@@ -12,6 +12,7 @@ import type { PluginFunction } from '@athoscommerce/snap-controller';
 import { pluginBackgroundFilters as shopifyPluginBackgroundFilters } from './library/plugins/shopify/pluginBackgroundFilters';
 import { pluginMutateResults as shopifyPluginMutateResults } from './library/plugins/shopify/pluginMutateResults';
 import { pluginAddToCart as shopifyPluginAddToCart } from './library/plugins/shopify/pluginAddToCart';
+import { pluginMarkets as shopifyPluginMarkets } from './library/plugins/shopify/pluginMarkets';
 import { pluginAddToCart as bigCommercePluginAddToCart } from './library/plugins/bigCommerce/pluginAddToCart';
 import { pluginAddToCart as magento2PluginAddToCart } from './library/plugins/magento2/pluginAddToCart';
 import { pluginAddToCart as commonPluginAddToCart } from './library/plugins/common/pluginAddToCart';
@@ -21,6 +22,7 @@ import { pluginBase as magento2PluginBase } from './library/plugins/magento2/plu
 import { pluginBackgroundFilters } from './library/plugins/common/pluginBackgroundFilters';
 import { pluginScrollToTop } from './library/plugins/common/pluginScrollToTop';
 import { pluginLogger } from './library/plugins/common/pluginLogger';
+import { pluginKlaviyoEvents } from './library/plugins/common/pluginKlaviyoEvents';
 import { CustomComponent } from './library/components/CustomComponent';
 
 type LibraryComponentImport = {
@@ -46,6 +48,7 @@ export type LibraryImports = {
 			backgroundFilters: typeof shopifyPluginBackgroundFilters;
 			mutateResults: typeof shopifyPluginMutateResults;
 			addToCart: typeof shopifyPluginAddToCart;
+			markets: typeof shopifyPluginMarkets;
 		};
 		bigcommerce: {
 			backgroundFilters: PluginFunction;
@@ -61,6 +64,7 @@ export type LibraryImports = {
 			scrollToTop: typeof pluginScrollToTop;
 			logger: typeof pluginLogger;
 			addToCart: typeof commonPluginAddToCart;
+			klaviyoEvents: typeof pluginKlaviyoEvents;
 		};
 		custom?: {
 			[name: string]: PluginFunction;
@@ -152,7 +156,6 @@ export type LibraryImports = {
 		results: LibraryComponentImport;
 		searchHeader: LibraryComponentImport;
 		sidebar: LibraryComponentImport;
-		mobileSidebar: LibraryComponentImport;
 		toolbar: LibraryComponentImport;
 		termsList: LibraryComponentImport;
 	};
@@ -244,7 +247,6 @@ export class LibraryStore {
 		results: LibraryComponentMap;
 		searchHeader: LibraryComponentMap;
 		sidebar: LibraryComponentMap;
-		mobileSidebar: LibraryComponentMap;
 		toolbar: LibraryComponentMap;
 		termsList: LibraryComponentMap;
 	} = {
@@ -315,7 +317,6 @@ export class LibraryStore {
 		results: {},
 		searchHeader: {},
 		sidebar: {},
-		mobileSidebar: {},
 		toolbar: {},
 		termsList: {},
 	};
@@ -361,6 +362,7 @@ export class LibraryStore {
 				backgroundFilters: shopifyPluginBackgroundFilters,
 				mutateResults: shopifyPluginMutateResults,
 				addToCart: shopifyPluginAddToCart,
+				markets: shopifyPluginMarkets,
 			},
 			bigcommerce: {
 				backgroundFilters: bigCommercePluginBackgroundFilters,
@@ -376,6 +378,7 @@ export class LibraryStore {
 				scrollToTop: pluginScrollToTop,
 				logger: pluginLogger,
 				addToCart: commonPluginAddToCart,
+				klaviyoEvents: pluginKlaviyoEvents,
 			},
 		},
 		component: {
@@ -552,7 +555,6 @@ export class LibraryStore {
 			results: {},
 			searchHeader: {},
 			sidebar: {},
-			mobileSidebar: {},
 			toolbar: {},
 			termsList: {},
 		},